Stop bundling the whole Redux Form library in the 'build' chunk
This patch is an alternative (and much simpler) way how to implement the changes
attempted in #21203.

From the Redux Form library, only the reducer is used in the `build` chunk. That's why
we import it there directly from the `redux-form/es/reducer` submodule.

Everything else (selectors, React components, ...) will be moved into the chunks that
use these parts of the library and should bear the costs.